🌱 Extract inline modals from god components
- OPAPolicies.tsx (2,288→913 lines): Extract PolicyDetailModal,
  ClusterOPAModal, and CreatePolicyModal into cards/opa/ subdirectory
  with shared types.ts and barrel index.ts
- GPUReservations.tsx (1,906→1,456 lines): Extract ReservationFormModal
  (450 lines) into gpu/ReservationFormModal.tsx
- Clusters.tsx (2,108→772 lines): Remove 1,336 lines of dead code
  (formatMetadata, ResourceDetailModal, inline NamespaceResources,
  _ClusterDetail) — all superseded by extracted components

No behavior changes. Build and lint pass.
