+++ web/src/components/cards/OPAPolicies.tsx
@@ -1,7 +1,6 @@
 import { useState, useEffect, useCallback, useRef, useMemo } from 'react'
-import { Shield, AlertTriangle, CheckCircle, ExternalLink, XCircle, Info, ChevronRight, RefreshCw, Plus, Edit3, Trash2, FileCode, LayoutTemplate, Sparkles, Copy, WifiOff, MessageSquareText, ScanSearch, Loader2 } from 'lucide-react'
+import { AlertTriangle, CheckCircle, ExternalLink, XCircle, Info, ChevronRight, RefreshCw, Plus, WifiOff, Loader2 } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
-import { BaseModal } from '../../lib/modals'
 import { useCardData, commonComparators } from '../../lib/cards/cardHooks'
 import { CardSearchInput, CardControlsRow, CardPaginationFooter, CardSkeleton } from '../../lib/cards/CardComponents'
 import { useClusters } from '../../hooks/useMCP'
@@ -10,8 +9,9 @@ import { kubectlProxy } from '../../lib/kubectlProxy'
 import { useCardLoadingState, useCardDemoState } from './CardDataContext'
 import { isDemoMode as checkIsDemoMode } from '../../lib/demoMode'
 import { DynamicCardErrorBoundary } from './DynamicCardErrorBoundary'
-import { useToast } from '../ui/Toast'
 import { LOCAL_AGENT_HTTP_URL, STORAGE_KEY_OPA_CACHE, STORAGE_KEY_OPA_CACHE_TIME } from '../../lib/constants'
+import { PolicyDetailModal, ClusterOPAModal, CreatePolicyModal } from './opa'
+import type { Policy, GatekeeperStatus, OPAClusterItem } from './opa'
 
 /** Cache TTL: 60 minutes — OPA installation status rarely changes */
 const CACHE_TTL_MS = 60 * 60 * 1000
@@ -25,49 +25,10 @@ const SORT_OPTIONS = [
   { value: 'policies' as const, label: 'Policies' },
 ]
 
-// Violation detail interface
-interface Violation {
-  name: string
-  namespace: string
-  kind: string
-  policy: string
-  message: string
-  severity: 'critical' | 'warning' | 'info'
-}
-
-// Policy interface for real data
-interface Policy {
-  name: string
-  kind: string // ConstraintTemplate kind
-  violations: number
-  mode: 'warn' | 'enforce' | 'dryrun' | 'deny'
-}
-
 interface OPAPoliciesProps {
   config?: {
     cluster?: string
   }
-}
-
-interface GatekeeperStatus {
-  cluster: string
-  installed: boolean
-  policyCount?: number
-  violationCount?: number
-  mode?: 'dryrun' | 'warn' | 'enforce' | 'deny'
-  modes?: ('warn' | 'enforce' | 'dryrun')[]  // All active modes for multi-badge display
-  loading: boolean
-  error?: string
-  policies?: Policy[]
-  violations?: Violation[]
-}
-
-// Item type for useCardData - enriched cluster with a 'cluster' field for filtering
-interface OPAClusterItem {
-  name: string
-  cluster: string // same as name, required for useCardData cluster filtering
-  healthy?: boolean
-  reachable?: boolean
 }
 
 /** Generate demo OPA statuses for instant display (no waiting for effects/API) */
@@ -89,176 +50,6 @@ function generateDemoStatuses(): Record<string, GatekeeperStatus> {
   }
   return result
 }
-
-// Common OPA Gatekeeper policy templates
-const POLICY_TEMPLATES = [
-  {
-    name: 'Require Labels',
-    description: 'Require specific labels on resources',
-    kind: 'K8sRequiredLabels',
-    template: `apiVersion: templates.gatekeeper.sh/v1
-kind: ConstraintTemplate
-metadata:
-  name: k8srequiredlabels
-spec:
-  crd:
-    spec:
-      names:
-        kind: K8sRequiredLabels
-      validation:
-        openAPIV3Schema:
-          type: object
-          properties:
-            labels:
-              type: array
-              items:
-                type: string
-  targets:
-    - target: admission.k8s.gatekeeper.sh
-      rego: |
-        package k8srequiredlabels
-        violation[{"msg": msg}] {
-          provided := {label | input.review.object.metadata.labels[label]}
-          required := {label | label := input.parameters.labels[_]}
-          missing := required - provided
-          count(missing) > 0
-          msg := sprintf("Missing required labels: %v", [missing])
-        }
----
-apiVersion: constraints.gatekeeper.sh/v1beta1
-kind: K8sRequiredLabels
-metadata:
-  name: require-team-label
-spec:
-  enforcementAction: warn
-  match:
-    kinds:
-      - apiGroups: [""]
-        kinds: ["Namespace"]
-  parameters:
-    labels: ["team", "owner"]`,
-  },
-  {
-    name: 'Restrict Image Registries',
-    description: 'Only allow images from approved registries',
-    kind: 'K8sAllowedRepos',
-    template: `apiVersion: templates.gatekeeper.sh/v1
-kind: ConstraintTemplate
-metadata:
-  name: k8sallowedrepos
-spec:
-  crd:
-    spec:
-      names:
-        kind: K8sAllowedRepos
-      validation:
-        openAPIV3Schema:
-          type: object
-          properties:
-            repos:
-              type: array
-              items:
-                type: string
-  targets:
-    - target: admission.k8s.gatekeeper.sh
-      rego: |
-        package k8sallowedrepos
-        violation[{"msg": msg}] {
-          container := input.review.object.spec.containers[_]
-          satisfied := [good | repo = input.parameters.repos[_]; good = startswith(container.image, repo)]
-          not any(satisfied)
-          msg := sprintf("Container image %v is not from an allowed registry", [container.image])
-        }
----
-apiVersion: constraints.gatekeeper.sh/v1beta1
-kind: K8sAllowedRepos
-metadata:
-  name: allowed-repos
-spec:
-  enforcementAction: warn
-  match:
-    kinds:
-      - apiGroups: [""]
-        kinds: ["Pod"]
-  parameters:
-    repos:
-      - "gcr.io/"
-      - "docker.io/"`,
-  },
-  {
-    name: 'Require Resource Limits',
-    description: 'Require CPU and memory limits on containers',
-    kind: 'K8sRequireResourceLimits',
-    template: `apiVersion: templates.gatekeeper.sh/v1
-kind: ConstraintTemplate
-metadata:
-  name: k8srequireresourcelimits
-spec:
-  crd:
-    spec:
-      names:
-        kind: K8sRequireResourceLimits
-  targets:
-    - target: admission.k8s.gatekeeper.sh
-      rego: |
-        package k8srequireresourcelimits
-        violation[{"msg": msg}] {
-          container := input.review.object.spec.containers[_]
-          not container.resources.limits.cpu
-          msg := sprintf("Container %v does not have CPU limits", [container.name])
-        }
-        violation[{"msg": msg}] {
-          container := input.review.object.spec.containers[_]
-          not container.resources.limits.memory
-          msg := sprintf("Container %v does not have memory limits", [container.name])
-        }
----
-apiVersion: constraints.gatekeeper.sh/v1beta1
-kind: K8sRequireResourceLimits
-metadata:
-  name: require-resource-limits
-spec:
-  enforcementAction: warn
-  match:
-    kinds:
-      - apiGroups: [""]
-        kinds: ["Pod"]`,
-  },
-  {
-    name: 'Block Privileged Containers',
-    description: 'Prevent privileged containers from running',
-    kind: 'K8sBlockPrivileged',
-    template: `apiVersion: templates.gatekeeper.sh/v1
-kind: ConstraintTemplate
-metadata:
-  name: k8sblockprivileged
-spec:
-  crd:
-    spec:
-      names:
-        kind: K8sBlockPrivileged
-  targets:
-    - target: admission.k8s.gatekeeper.sh
-      rego: |
-        package k8sblockprivileged
-        violation[{"msg": msg}] {
-          container := input.review.object.spec.containers[_]
-          container.securityContext.privileged == true
-          msg := sprintf("Privileged containers are not allowed: %v", [container.name])
-        }
----
-apiVersion: constraints.gatekeeper.sh/v1beta1
-kind: K8sBlockPrivileged
-metadata:
-  name: block-privileged
-spec:
-  enforcementAction: deny
-  match:
-    kinds:
-      - apiGroups: [""]
-        kinds: ["Pod"]`,
-  },
-]
 
 // Module-level flag to prevent StrictMode double-checks
 // This persists across component mounts within the same page load
@@ -329,7 +120,7 @@ async function checkGatekeeperDetails(clusterName: string): Promise<GatekeeperSt
     }
 
     // Fetch sample violations (only if there are violations to show)
-    const violations: Violation[] = []
+    const violations: GatekeeperStatus['violations'] = []
     if (totalViolations > 0 && policies.length > 0) {
       const policyWithViolations = policies.find(p => p.violations > 0)
       if (policyWithViolations) {
@@ -375,1164 +166,6 @@ async function checkGatekeeperDetails(clusterName: string): Promise<GatekeeperSt
     // Details failed but Gatekeeper is installed — show installed with no details
     return { cluster: clusterName, installed: true, loading: false, policyCount: 0, violationCount: 0 }
   }
-}
-
-
-// Policy Detail Modal
-function PolicyDetailModal({
-  isOpen,
-  onClose,
-  policy,
-  violations,
-  onAddPolicy
-}: {
-  isOpen: boolean
-  onClose: () => void
-  policy: Policy
-  violations: Violation[]
-  onAddPolicy: () => void
-}) {
-  const { t } = useTranslation(['cards', 'common'])
-  // Get violations for this policy
-  const policyViolations = violations.filter(v => v.policy === policy.name)
-
-  const getModeColor = (mode: string) => {
-    switch (mode) {
-      case 'enforce':
-      case 'deny':
-        return 'text-red-400 bg-red-500/20'
-      case 'warn': return 'text-amber-400 bg-amber-500/20'
-      default: return 'text-blue-400 bg-blue-500/20'
-    }
-  }
-
-  return (
-    <BaseModal isOpen={isOpen} onClose={onClose} size="md">
-      <BaseModal.Header
-        title={policy.name}
-        description={policy.kind}
-        icon={Shield}
-        onClose={onClose}
-        showBack={false}
-      />
-
-      <BaseModal.Content className="max-h-[50vh]">
-        {/* Policy Info */}
-        <div className="mb-4 pb-4 border-b border-border">
-          <div className="flex items-center gap-4">
-            <div className="flex-1">
-              <p className="text-xs text-muted-foreground mb-1">Enforcement Mode</p>
-              <span className={`px-2 py-1 rounded text-sm font-medium ${getModeColor(policy.mode)}`}>
-                {policy.mode}
-              </span>
-            </div>
-            <div className="flex-1">
-              <p className="text-xs text-muted-foreground mb-1">Violations</p>
-              <p className={`text-2xl font-bold ${policy.violations > 0 ? 'text-amber-400' : 'text-green-400'}`}>
-                {policy.violations}
-              </p>
-            </div>
-          </div>
-        </div>
-
-        {/* Violations List */}
-        {policyViolations.length === 0 ? (
-          <div className="text-center py-8 text-muted-foreground">
-            <CheckCircle className="w-8 h-8 mx-auto mb-2 text-green-400" />
-            <p>{t('messages.noViolations')}</p>
-          </div>
-        ) : (
-          <div className="space-y-2">
-            {policyViolations.map((violation, idx) => (
-              <div
-                key={idx}
-                className="p-3 rounded-lg bg-secondary/30 hover:bg-secondary/50 transition-colors"
-              >
-                <div className="flex items-start justify-between mb-1">
-                  <span className="text-sm font-medium text-foreground">{violation.name}</span>
-                  <span className="text-xs text-muted-foreground">{violation.kind}</span>
-                </div>
-                <p className="text-sm text-muted-foreground mb-1">{violation.message}</p>
-                <span className="text-xs text-muted-foreground">Namespace: <span className="text-foreground">{violation.namespace}</span></span>
-              </div>
-            ))}
-          </div>
-        )}
-      </BaseModal.Content>
-
-      <BaseModal.Footer>
-        <a
-          href={`https://open-policy-agent.github.io/gatekeeper-library/website/${policy.kind.toLowerCase()}/`}
-          target="_blank"
-          rel="noopener noreferrer"
-          className="text-sm text-purple-400 hover:text-purple-300 flex items-center gap-1"
-        >
-          Policy Documentation
-          <ExternalLink className="w-3 h-3" />
-        </a>
-        <div className="flex-1" />
-        <div className="flex items-center gap-2">
-          <button
-            onClick={() => {
-              onClose()
-              onAddPolicy()
-            }}
-            className="px-4 py-2 bg-purple-500/20 text-purple-400 rounded-lg hover:bg-purple-500/30 transition-colors"
-          >
-            Create Similar Policy
-          </button>
-          <button
-            onClick={onClose}
-            className="px-4 py-2 bg-secondary text-foreground rounded-lg hover:bg-secondary/80 transition-colors"
-          >
-            Close
-          </button>
-        </div>
-      </BaseModal.Footer>
-    </BaseModal>
-  )
-}
-
-// Tab type for ClusterOPAModal
-type OPAModalTab = 'policies' | 'violations'
-
-// Cluster OPA Modal - Full CRUD for OPA policies
-function ClusterOPAModal({
-  isOpen,
-  onClose,
-  clusterName,
-  policies,
-  violations,
-  onRefresh,
-  startMission
-}: {
-  isOpen: boolean
-  onClose: () => void
-  clusterName: string
-  policies: Policy[]
-  violations: Violation[]
-  onRefresh: () => void
-  startMission: (mission: {
-    title: string
-    description: string
-    type: 'upgrade' | 'troubleshoot' | 'analyze' | 'deploy' | 'repair' | 'custom'
-    cluster: string
-    initialPrompt: string
-    context?: Record<string, unknown>
-  }) => void
-}) {
-  const { t } = useTranslation(['cards', 'common'])
-  const { showToast } = useToast()
-  const [activeTab, setActiveTab] = useState<OPAModalTab>('policies')
-  const [showCreateMenu, setShowCreateMenu] = useState(false)
-  const [showTemplateModal, setShowTemplateModal] = useState(false)
-  const [showYamlEditor, setShowYamlEditor] = useState(false)
-  const [editingPolicy, setEditingPolicy] = useState<Policy | null>(null)
-  const [yamlContent, setYamlContent] = useState('')
-  const [deleteConfirm, setDeleteConfirm] = useState<Policy | null>(null)
-  const createMenuRef = useRef<HTMLDivElement>(null)
-
-  // Close create menu when clicking outside
-  useEffect(() => {
-    const handleClickOutside = (e: MouseEvent) => {
-      if (createMenuRef.current && !createMenuRef.current.contains(e.target as Node)) {
-        setShowCreateMenu(false)
-      }
-    }
-    if (showCreateMenu) {
-      document.addEventListener('mousedown', handleClickOutside)
-    }
-    return () => document.removeEventListener('mousedown', handleClickOutside)
-  }, [showCreateMenu])
-
-  const severityCounts = {
-    critical: violations.filter(v => v.severity === 'critical').length,
-    warning: violations.filter(v => v.severity === 'warning').length,
-    info: violations.filter(v => v.severity === 'info').length,
-  }
-
-  const getSeverityColor = (severity: string) => {
-    switch (severity) {
-      case 'critical': return 'text-red-400 bg-red-500/20'
-      case 'warning': return 'text-amber-400 bg-amber-500/20'
-      default: return 'text-blue-400 bg-blue-500/20'
-    }
-  }
-
-  const getModeColor = (mode: string) => {
-    switch (mode) {
-      case 'enforce':
-      case 'deny':
-        return 'text-red-400 bg-red-500/20'
-      case 'warn': return 'text-amber-400 bg-amber-500/20'
-      default: return 'text-blue-400 bg-blue-500/20'
-    }
-  }
-
-  // Create policy with AI
-  const handleCreateWithAI = () => {
-    setShowCreateMenu(false)
-    onClose()
-    startMission({
-      title: 'Create OPA Gatekeeper Policy',
-      description: 'Create a new OPA Gatekeeper policy with AI assistance',
-      type: 'deploy',
-      cluster: clusterName,
-      initialPrompt: `I want to create a new OPA Gatekeeper policy for the cluster "${clusterName}".
-
-Please help me:
-1. Ask me what kind of policy I want to enforce (e.g., require labels, restrict images, enforce resource limits)
-2. Generate the appropriate ConstraintTemplate and Constraint
-3. Help me apply it to the cluster
-4. Test that the policy is working
-
-Let's start by discussing what kind of policy I need.`,
-      context: { clusterName },
-    })
-  }
-
-  // Use a template
-  const handleUseTemplate = (template: typeof POLICY_TEMPLATES[0]) => {
-    setYamlContent(template.template)
-    setEditingPolicy(null)
-    setShowTemplateModal(false)
-    setShowYamlEditor(true)
-  }
-
-  // Edit policy with AI
-  const handleEditWithAI = (policy: Policy) => {
-    onClose()
-    startMission({
-      title: `Edit Policy: ${policy.name}`,
-      description: `Modify OPA Gatekeeper policy ${policy.name}`,
-      type: 'deploy',
-      cluster: clusterName,
-      initialPrompt: `I want to edit the OPA Gatekeeper policy "${policy.name}" (kind: ${policy.kind}) on cluster "${clusterName}".
-
-Current enforcement mode: ${policy.mode}
-Current violations: ${policy.violations}
-
-Please help me:
-1. Fetch the current policy YAML
-2. Ask me what changes I want to make
-3. Update the policy
-4. Verify the changes
-
-What would you like to modify about this policy?`,
-      context: { clusterName, policy },
-    })
-  }
-
-  // Edit policy YAML directly
-  const handleEditYaml = async (policy: Policy) => {
-    setEditingPolicy(policy)
-    setYamlContent('# Loading policy YAML...\n# Fetching from cluster: ' + clusterName)
-    setShowYamlEditor(true)  // Show modal immediately
-
-    // Fetch the current YAML in background
-    const cmd = ['get', policy.kind.toLowerCase(), policy.name, '-o', 'yaml']
-
-    try {
-      // Use priority: true to bypass the queue for immediate execution (interactive user action)
-      const result = await kubectlProxy.exec(cmd, { context: clusterName, timeout: 30000, priority: true })
-
-      if (result.output && result.output.trim()) {
-        setYamlContent(result.output)
-      } else if (result.error) {
-        setYamlContent(`# Failed to fetch policy YAML\n# Error: ${result.error}\n\n# You can write new YAML here`)
-      } else {
-        setYamlContent('# No YAML returned from cluster\n# You can write new YAML here')
-      }
-    } catch (err) {
-      console.error('[OPA] Failed to fetch policy YAML:', err)
-      setYamlContent(`# Failed to fetch policy YAML\n# Error: ${err}\n\n# You can write new YAML here`)
-    }
-  }
-
-  // Apply YAML changes via AI (validates and applies safely)
-  const handleApplyYaml = () => {
-    const action = editingPolicy ? 'update' : 'create'
-    setShowYamlEditor(false)
-    onClose()
-    startMission({
-      title: editingPolicy ? `Apply Policy: ${editingPolicy.name}` : 'Apply OPA Policy',
-      description: `Apply OPA Gatekeeper policy YAML to ${clusterName}`,
-      type: 'deploy',
-      cluster: clusterName,
-      initialPrompt: `Please apply the following OPA Gatekeeper policy YAML to cluster "${clusterName}":
-
-\`\`\`yaml
-${yamlContent}
-\`\`\`
-
-Steps:
-1. Review the YAML for any issues
-2. Apply it to the cluster using kubectl apply
-3. Verify the policy was created/updated successfully
-4. Check if there are any immediate violations
-
-Please proceed with applying this policy.`,
-      context: { clusterName, action, yaml: yamlContent },
-    })
-    setYamlContent('')
-    setEditingPolicy(null)
-  }
-
-  // Toggle enforcement mode
-  const handleToggleMode = async (policy: Policy) => {
-    const newMode = policy.mode === 'enforce' ? 'warn' : policy.mode === 'warn' ? 'dryrun' : 'enforce'
-    try {
-      await kubectlProxy.exec(
-        ['patch', policy.kind.toLowerCase(), policy.name, '--type=merge', '-p', `{"spec":{"enforcementAction":"${newMode}"}}`],
-        { context: clusterName, timeout: 15000 }
-      )
-      showToast('Policy mode updated successfully', 'success')
-      onRefresh()
-    } catch (err) {
-      console.error('Failed to toggle mode:', err)
-      showToast('Failed to toggle policy mode', 'error')
-    }
-  }
-
-  // Delete policy
-  const handleDelete = async (policy: Policy) => {
-    try {
-      await kubectlProxy.exec(
-        ['delete', policy.kind.toLowerCase(), policy.name],
-        { context: clusterName, timeout: 15000 }
-      )
-      setDeleteConfirm(null)
-      showToast('Policy deleted successfully', 'success')
-      onRefresh()
-    } catch (err) {
-      console.error('Failed to delete policy:', err)
-      showToast('Failed to delete policy', 'error')
-    }
-  }
-
-  // Disable parent modal's Escape handler when a child modal is open
-  const hasChildModalOpen = showTemplateModal || showYamlEditor || !!deleteConfirm
-
-  return (
-    <>
-      <BaseModal isOpen={isOpen} onClose={onClose} size="lg" closeOnEscape={!hasChildModalOpen}>
-        <BaseModal.Header
-          title="OPA Gatekeeper"
-          description={clusterName}
-          icon={Shield}
-          onClose={onClose}
-          showBack={false}
-        />
-
-        <BaseModal.Content className="max-h-[60vh]">
-          {/* Tabs */}
-          <div className="flex items-center justify-between mb-4 pb-3 border-b border-border">
-            <div className="flex gap-1">
-              <button
-                onClick={() => setActiveTab('policies')}
-                className={`px-3 py-1.5 text-sm rounded-lg transition-colors ${
-                  activeTab === 'policies'
-                    ? 'bg-purple-500/20 text-purple-400'
-                    : 'text-muted-foreground hover:text-foreground hover:bg-secondary'
-                }`}
-              >
-                Policies ({policies.length})
-              </button>
-              <button
-                onClick={() => setActiveTab('violations')}
-                className={`px-3 py-1.5 text-sm rounded-lg transition-colors ${
-                  activeTab === 'violations'
-                    ? 'bg-purple-500/20 text-purple-400'
-                    : 'text-muted-foreground hover:text-foreground hover:bg-secondary'
-                }`}
-              >
-                Violations ({violations.length})
-              </button>
-            </div>
-
-            {/* Create Policy Button */}
-            <div ref={createMenuRef} className="relative">
-              <button
-                onClick={() => setShowCreateMenu(!showCreateMenu)}
-                className="flex items-center gap-1.5 px-3 py-1.5 text-sm bg-purple-500/20 text-purple-400 rounded-lg hover:bg-purple-500/30 transition-colors"
-              >
-                <Plus className="w-4 h-4" />
-                Create Policy
-              </button>
-              {showCreateMenu && (
-                <div className="absolute right-0 top-full mt-1 w-56 bg-card border border-border rounded-lg shadow-lg z-50 py-1">
-                  <button
-                    onClick={handleCreateWithAI}
-                    className="w-full px-3 py-2 text-left text-sm hover:bg-secondary transition-colors flex items-center gap-2"
-                  >
-                    <Sparkles className="w-4 h-4 text-purple-400" />
-                    <div>
-                      <div className="font-medium">Create with AI</div>
-                      <div className="text-xs text-muted-foreground">AI-assisted policy creation</div>
-                    </div>
-                  </button>
-                  <button
-                    onClick={() => { setShowCreateMenu(false); setShowTemplateModal(true) }}
-                    className="w-full px-3 py-2 text-left text-sm hover:bg-secondary transition-colors flex items-center gap-2"
-                  >
-                    <LayoutTemplate className="w-4 h-4 text-blue-400" />
-                    <div>
-                      <div className="font-medium">From Template</div>
-                      <div className="text-xs text-muted-foreground">Use a pre-built policy</div>
-                    </div>
-                  </button>
-                  <button
-                    onClick={() => { setShowCreateMenu(false); setYamlContent(''); setEditingPolicy(null); setShowYamlEditor(true) }}
-                    className="w-full px-3 py-2 text-left text-sm hover:bg-secondary transition-colors flex items-center gap-2"
-                  >
-                    <FileCode className="w-4 h-4 text-green-400" />
-                    <div>
-                      <div className="font-medium">Custom YAML</div>
-                      <div className="text-xs text-muted-foreground">Write policy manually</div>
-                    </div>
-                  </button>
-                </div>
-              )}
-            </div>
-          </div>
-
-          {/* Policies Tab */}
-          {activeTab === 'policies' && (
-            <div className="space-y-2">
-              {policies.length === 0 ? (
-                <div className="text-center py-8 text-muted-foreground">
-                  <Shield className="w-8 h-8 mx-auto mb-2 opacity-50" />
-                  <p>{t('messages.noPoliciesConfigured')}</p>
-                  <p className="text-xs mt-1">{t('messages.createPolicyPrompt')}</p>
-                </div>
-              ) : (
-                policies.map(policy => (
-                  <div
-                    key={policy.name}
-                    onClick={() => handleEditYaml(policy)}
-                    className="p-3 rounded-lg bg-secondary/30 hover:bg-secondary/50 transition-colors cursor-pointer group"
-                  >
-                    <div className="flex items-center justify-between mb-2">
-                      <div className="flex items-center gap-2">
-                        <span className="text-sm font-medium text-foreground group-hover:text-purple-400 transition-colors">{policy.name}</span>
-                        <span className="text-xs text-muted-foreground">({policy.kind})</span>
-                      </div>
-                      <div className="flex items-center gap-2">
-                        <button
-                          onClick={(e) => { e.stopPropagation(); handleToggleMode(policy) }}
-                          className={`px-2 py-0.5 rounded text-xs font-medium transition-colors hover:opacity-80 ${getModeColor(policy.mode)}`}
-                          title="Click to cycle: enforce → warn → dryrun"
-                        >
-                          {policy.mode}
-                        </button>
-                      </div>
-                    </div>
-                    <div className="flex items-center justify-between">
-                      <div className="flex items-center gap-3 text-xs">
-                        {policy.violations > 0 ? (
-                          <span className="flex items-center gap-1 text-amber-400">
-                            <AlertTriangle className="w-3 h-3" />
-                            {policy.violations} violations
-                          </span>
-                        ) : (
-                          <span className="flex items-center gap-1 text-green-400">
-                            <CheckCircle className="w-3 h-3" />
-                            No violations
-                          </span>
-                        )}
-                      </div>
-                      <div className="flex items-center gap-1">
-                        <button
-                          onClick={(e) => { e.stopPropagation(); handleEditWithAI(policy) }}
-                          className="p-1.5 rounded hover:bg-secondary text-purple-400 transition-colors"
-                          title="Edit with AI"
-                        >
-                          <Sparkles className="w-3.5 h-3.5" />
-                        </button>
-                        <button
-                          onClick={(e) => { e.stopPropagation(); handleEditYaml(policy) }}
-                          className="p-1.5 rounded hover:bg-secondary text-muted-foreground hover:text-foreground transition-colors"
-                          title="Edit YAML"
-                        >
-                          <Edit3 className="w-3.5 h-3.5" />
-                        </button>
-                        <button
-                          onClick={(e) => { e.stopPropagation(); setDeleteConfirm(policy) }}
-                          className="p-1.5 rounded hover:bg-red-500/20 text-muted-foreground hover:text-red-400 transition-colors"
-                          title="Delete policy"
-                        >
-                          <Trash2 className="w-3.5 h-3.5" />
-                        </button>
-                      </div>
-                    </div>
-                  </div>
-                ))
-              )}
-            </div>
-          )}
-
-          {/* Violations Tab */}
-          {activeTab === 'violations' && (
-            <>
-              {/* Summary */}
-              <div className="grid grid-cols-3 gap-3 mb-4 pb-4 border-b border-border">
-                <div className="p-3 rounded-lg bg-red-500/10 border border-red-500/20 text-center">
-                  <p className="text-2xl font-bold text-red-400">{severityCounts.critical}</p>
-                  <p className="text-xs text-muted-foreground">{t('common:common.critical')}</p>
-                </div>
-                <div className="p-3 rounded-lg bg-amber-500/10 border border-amber-500/20 text-center">
-                  <p className="text-2xl font-bold text-amber-400">{severityCounts.warning}</p>
-                  <p className="text-xs text-muted-foreground">{t('common:common.warning')}</p>
-                </div>
-                <div className="p-3 rounded-lg bg-blue-500/10 border border-blue-500/20 text-center">
-                  <p className="text-2xl font-bold text-blue-400">{severityCounts.info}</p>
-                  <p className="text-xs text-muted-foreground">Info</p>
-                </div>
-              </div>
-
-              {/* Violations List */}
-              <div className="space-y-2">
-                {violations.length === 0 ? (
-                  <div className="text-center py-8 text-muted-foreground">
-                    <CheckCircle className="w-8 h-8 mx-auto mb-2 text-green-400" />
-                    <p className="text-green-400">No violations</p>
-                    <p className="text-xs mt-1">All resources comply with policies</p>
-                  </div>
-                ) : (
-                  [...violations]
-                    .sort((a, b) => {
-                      const severityOrder = { critical: 0, warning: 1, info: 2 }
-                      return severityOrder[a.severity] - severityOrder[b.severity]
-                    })
-                    .map((violation, idx) => (
-                    <div
-                      key={idx}
-                      className="p-3 rounded-lg bg-secondary/30 hover:bg-secondary/50 transition-colors"
-                    >
-                      <div className="flex items-start justify-between mb-2">
-                        <div className="flex items-center gap-2">
-                          <span className={`px-2 py-0.5 rounded text-xs font-medium ${getSeverityColor(violation.severity)}`}>
-                            {violation.severity}
-                          </span>
-                          <span className="text-sm font-medium text-foreground">{violation.name}</span>
-                        </div>
-                        <span className="text-xs text-muted-foreground">{violation.kind}</span>
-                      </div>
-                      <p className="text-sm text-muted-foreground mb-2">{violation.message}</p>
-                      <div className="flex items-center gap-3 text-xs text-muted-foreground">
-                        <span>Namespace: <span className="text-foreground">{violation.namespace}</span></span>
-                        <span>Policy: <span className="text-orange-400">{violation.policy}</span></span>
-                      </div>
-                    </div>
-                  ))
-                )}
-              </div>
-            </>
-          )}
-        </BaseModal.Content>
-
-        <BaseModal.Footer>
-          <a
-            href="https://open-policy-agent.github.io/gatekeeper/website/docs/"
-            target="_blank"
-            rel="noopener noreferrer"
-            className="text-sm text-purple-400 hover:text-purple-300 flex items-center gap-1"
-          >
-            Documentation
-            <ExternalLink className="w-3 h-3" />
-          </a>
-          <div className="flex-1" />
-          <button
-            onClick={onClose}
-            className="px-4 py-2 bg-secondary text-foreground rounded-lg hover:bg-secondary/80 transition-colors"
-          >
-            Close
-          </button>
-        </BaseModal.Footer>
-      </BaseModal>
-
-      {/* Template Selection Modal */}
-      <BaseModal isOpen={showTemplateModal} onClose={() => setShowTemplateModal(false)} size="md">
-        <BaseModal.Header
-          title="Policy Templates"
-          description="Choose a template to start with"
-          icon={LayoutTemplate}
-          onClose={() => setShowTemplateModal(false)}
-          showBack={false}
-        />
-        <BaseModal.Content className="max-h-[50vh]">
-          <div className="space-y-2">
-            {POLICY_TEMPLATES.map(template => (
-              <button
-                key={template.name}
-                onClick={() => handleUseTemplate(template)}
-                className="w-full p-3 rounded-lg bg-secondary/30 hover:bg-secondary/50 transition-colors text-left"
-              >
-                <div className="flex items-center justify-between mb-1">
-                  <span className="text-sm font-medium text-foreground">{template.name}</span>
-                  <span className="text-xs text-muted-foreground">{template.kind}</span>
-                </div>
-                <p className="text-xs text-muted-foreground">{template.description}</p>
-              </button>
-            ))}
-          </div>
-        </BaseModal.Content>
-      </BaseModal>
-
-      {/* YAML Editor Modal */}
-      <BaseModal isOpen={showYamlEditor} onClose={() => setShowYamlEditor(false)} size="lg">
-        <BaseModal.Header
-          title={editingPolicy ? `Edit: ${editingPolicy.name}` : 'Create Policy'}
-          description="Edit the YAML and apply to cluster"
-          icon={FileCode}
-          onClose={() => setShowYamlEditor(false)}
-          showBack={false}
-        />
-        <BaseModal.Content className="!overflow-visible">
-          <div className="space-y-3">
-            <div className="flex items-center justify-between text-xs">
-              <span className="text-muted-foreground">YAML will be applied to: <span className="text-foreground">{clusterName}</span></span>
-              <button
-                onClick={() => navigator.clipboard.writeText(yamlContent)}
-                className="flex items-center gap-1 text-muted-foreground hover:text-foreground transition-colors"
-              >
-                <Copy className="w-3 h-3" />
-                Copy
-              </button>
-            </div>
-            <textarea
-              value={yamlContent}
-              onChange={(e) => setYamlContent(e.target.value)}
-              className="w-full h-[60vh] p-3 bg-secondary/50 border border-border rounded-lg font-mono text-sm text-foreground resize-none focus:outline-none focus:ring-1 focus:ring-purple-500/50"
-              placeholder="# Paste or write your ConstraintTemplate and Constraint YAML here..."
-              spellCheck={false}
-            />
-          </div>
-        </BaseModal.Content>
-        <BaseModal.Footer>
-          <button
-            onClick={() => setShowYamlEditor(false)}
-            className="px-4 py-2 text-muted-foreground hover:text-foreground transition-colors"
-          >
-            Cancel
-          </button>
-          <div className="flex-1" />
-          <button
-            onClick={handleApplyYaml}
-            disabled={!yamlContent.trim() || yamlContent.startsWith('# Loading')}
-            className="px-4 py-2 bg-purple-500 text-white rounded-lg hover:bg-purple-600 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
-          >
-            Apply
-          </button>
-        </BaseModal.Footer>
-      </BaseModal>
-
-      {/* Delete Confirmation Modal */}
-      <BaseModal isOpen={!!deleteConfirm} onClose={() => setDeleteConfirm(null)} size="sm">
-        <BaseModal.Header
-          title="Delete Policy"
-          description="This action cannot be undone"
-          icon={Trash2}
-          onClose={() => setDeleteConfirm(null)}
-          showBack={false}
-        />
-        <BaseModal.Content>
-          <div className="space-y-3">
-            <p className="text-sm text-muted-foreground">
-              Are you sure you want to delete the policy <span className="text-foreground font-medium">{deleteConfirm?.name}</span>?
-            </p>
-            {deleteConfirm && deleteConfirm.violations > 0 && (
-              <div className="p-3 rounded-lg bg-amber-500/10 border border-amber-500/30 text-sm">
-                <div className="flex items-center gap-2 text-amber-400 mb-1">
-                  <AlertTriangle className="w-4 h-4" />
-                  <span className="font-medium">{t('common:common.warning')}</span>
-                </div>
-                <p className="text-muted-foreground">
-                  This policy has {deleteConfirm.violations} active violations that will be cleared.
-                </p>
-              </div>
-            )}
-          </div>
-        </BaseModal.Content>
-        <BaseModal.Footer>
-          <button
-            onClick={() => setDeleteConfirm(null)}
-            className="px-4 py-2 text-muted-foreground hover:text-foreground transition-colors"
-          >
-            Cancel
-          </button>
-          <div className="flex-1" />
-          <button
-            onClick={() => deleteConfirm && handleDelete(deleteConfirm)}
-            className="flex items-center gap-2 px-4 py-2 bg-red-500/20 text-red-400 rounded-lg hover:bg-red-500/30 transition-colors"
-          >
-            <Trash2 className="w-4 h-4" />
-            Delete Policy
-          </button>
-        </BaseModal.Footer>
-      </BaseModal>
-    </>
-  )
-}
-
-// Creation flow type for CreatePolicyModal
-type CreateFlow = 'choose' | 'describe' | 'template' | 'yaml'
-
-// CreatePolicyModal — AI-driven policy creation from the main card
-function CreatePolicyModal({
-  isOpen,
-  onClose,
-  statuses,
-  startMission,
-}: {
-  isOpen: boolean
-  onClose: () => void
-  statuses: Record<string, GatekeeperStatus>
-  startMission: (mission: {
-    title: string
-    description: string
-    type: 'upgrade' | 'troubleshoot' | 'analyze' | 'deploy' | 'repair' | 'custom'
-    cluster: string
-    initialPrompt: string
-    context?: Record<string, unknown>
-  }) => void
-}) {
-  const { showToast } = useToast()
-  const [selectedCluster, setSelectedCluster] = useState('')
-  const [flow, setFlow] = useState<CreateFlow>('choose')
-  const [userDescription, setUserDescription] = useState('')
-  const [yamlContent, setYamlContent] = useState('')
-  const [isAnalyzing, setIsAnalyzing] = useState(false)
-  const mountedRef = useRef(true)
-  useEffect(() => () => { mountedRef.current = false }, [])
-
-  // Clusters that have Gatekeeper installed
-  const installedClusters = useMemo(
-    () => Object.entries(statuses).filter(([, s]) => s.installed).map(([name]) => name),
-    [statuses]
-  )
-
-  // Reset state and auto-select first installed cluster when modal opens
-  useEffect(() => {
-    if (isOpen) {
-      setFlow('choose')
-      setUserDescription('')
-      setYamlContent('')
-      setIsAnalyzing(false)
-      setSelectedCluster(installedClusters.length > 0 ? installedClusters[0] : '')
-    }
-  }, [isOpen, installedClusters])
-
-  // Gather cluster security data and start AI mission
-  const handleAnalyzeAndSuggest = async () => {
-    if (!selectedCluster) return
-    setIsAnalyzing(true)
-
-    try {
-      // Gather security data from the cluster
-      let securitySummary = ''
-      let resourceLimitsSummary = ''
-
-      try {
-        const podsResult = await kubectlProxy.exec(
-          ['get', 'pods', '-A', '-o', 'json'],
-          { context: selectedCluster, timeout: 20000 }
-        )
-
-        if (podsResult.output) {
-          const podsData = JSON.parse(podsResult.output)
-          const pods = podsData.items || []
-
-          // Analyze security issues
-          let privilegedCount = 0
-          let hostNetworkCount = 0
-          let runAsRootCount = 0
-          let noLimitsCount = 0
-          const issueDetails: string[] = []
-
-          for (const pod of pods) {
-            const ns = pod.metadata?.namespace || 'unknown'
-            // Skip system namespaces
-            if (ns.startsWith('kube-') || ns === 'gatekeeper-system') continue
-
-            if (pod.spec?.hostNetwork) {
-              hostNetworkCount++
-              issueDetails.push(`- Pod ${ns}/${pod.metadata?.name}: uses hostNetwork`)
-            }
-
-            for (const container of (pod.spec?.containers || [])) {
-              if (container.securityContext?.privileged) {
-                privilegedCount++
-                issueDetails.push(`- Container ${container.name} in ${ns}/${pod.metadata?.name}: runs privileged`)
-              }
-              if (container.securityContext?.runAsUser === 0 ||
-                  (!container.securityContext?.runAsNonRoot && !pod.spec?.securityContext?.runAsNonRoot)) {
-                runAsRootCount++
-              }
-              if (!container.resources?.limits?.cpu || !container.resources?.limits?.memory) {
-                noLimitsCount++
-              }
-            }
-          }
-
-          securitySummary = [
-            `Privileged containers: ${privilegedCount}`,
-            `Host network pods: ${hostNetworkCount}`,
-            `Containers potentially running as root: ${runAsRootCount}`,
-            ...(issueDetails.length > 0 ? ['', 'Details (first 10):', ...issueDetails.slice(0, 10)] : []),
-          ].join('\n')
-
-          resourceLimitsSummary = `Containers without CPU/memory limits: ${noLimitsCount}`
-        }
-      } catch {
-        securitySummary = 'Could not fetch pod data (cluster may be unreachable)'
-        resourceLimitsSummary = 'N/A'
-      }
-
-      // Get existing policies
-      const existingPolicies = (statuses[selectedCluster]?.policies ?? []).map(p => p.name).join(', ') || 'none'
-
-      if (!mountedRef.current) return
-      onClose()
-      startMission({
-        title: `AI: Analyze & Create Policies for ${selectedCluster}`,
-        description: 'AI scans cluster security posture and suggests OPA policies',
-        type: 'deploy',
-        cluster: selectedCluster,
-        initialPrompt: `Analyze the security posture of cluster "${selectedCluster}" and create OPA Gatekeeper policies to address the gaps.
-
-Current cluster state:
---- Security Issues ---
-${securitySummary || 'No issues detected'}
-
---- Resource Limits ---
-${resourceLimitsSummary || 'All containers have limits'}
-
---- Existing OPA Policies ---
-${existingPolicies}
-
-Based on this analysis:
-1. Identify the most critical gaps that OPA policies should address
-2. For each gap, generate a ConstraintTemplate + Constraint YAML
-3. Start with enforcementAction: dryrun so existing workloads aren't disrupted
-4. Apply the policies to the cluster with my approval
-5. After applying, check for immediate violations and report them
-
-Start with the highest-priority policy.`,
-        context: { cluster: selectedCluster },
-      })
-    } catch (err) {
-      console.error('[OPA] Failed to analyze cluster:', err)
-      if (mountedRef.current) showToast('Failed to gather cluster data', 'error')
-    } finally {
-      if (mountedRef.current) setIsAnalyzing(false)
-    }
-  }
-
-  // Start AI mission with user's description
-  const handleDescribeMission = () => {
-    if (!selectedCluster || !userDescription.trim()) return
-    onClose()
-    startMission({
-      title: `AI: Create Policy for ${selectedCluster}`,
-      description: 'AI generates OPA policy from user description',
-      type: 'deploy',
-      cluster: selectedCluster,
-      initialPrompt: `Create an OPA Gatekeeper policy for cluster "${selectedCluster}" based on this requirement:
-
-"${userDescription.trim()}"
-
-Please:
-1. Generate the ConstraintTemplate and Constraint YAML
-2. Explain what the policy does and what it catches
-3. Apply it to the cluster with enforcementAction: dryrun first
-4. Show any immediate violations
-5. Ask if I want to escalate to warn or enforce`,
-      context: { cluster: selectedCluster, description: userDescription.trim() },
-    })
-  }
-
-  // Apply custom YAML via AI mission
-  const handleApplyCustomYaml = () => {
-    if (!selectedCluster || !yamlContent.trim()) return
-    onClose()
-    startMission({
-      title: 'Apply OPA Policy',
-      description: `Apply OPA Gatekeeper policy YAML to ${selectedCluster}`,
-      type: 'deploy',
-      cluster: selectedCluster,
-      initialPrompt: `Please apply the following OPA Gatekeeper policy YAML to cluster "${selectedCluster}":
-
-\`\`\`yaml
-${yamlContent}
-\`\`\`
-
-Steps:
-1. Review the YAML for any issues
-2. Apply it to the cluster using kubectl apply
-3. Verify the policy was created/updated successfully
-4. Check if there are any immediate violations
-
-Please proceed with applying this policy.`,
-      context: { cluster: selectedCluster, yaml: yamlContent },
-    })
-  }
-
-  // Use a template — populate YAML editor
-  const handleUseTemplate = (template: typeof POLICY_TEMPLATES[0]) => {
-    setYamlContent(template.template)
-    setFlow('yaml')
-  }
-
-  // No installed clusters — empty state
-  const noGatekeeper = installedClusters.length === 0
-
-  return (
-    <BaseModal isOpen={isOpen} onClose={onClose} size="md" closeOnEscape={flow === 'choose'}>
-      <BaseModal.Header
-        title="Create OPA Policy"
-        description="AI-powered policy creation"
-        icon={Shield}
-        onClose={onClose}
-        showBack={flow !== 'choose'}
-        onBack={() => setFlow('choose')}
-      />
-
-      <BaseModal.Content className="max-h-[60vh]">
-        {noGatekeeper ? (
-          /* No clusters with Gatekeeper */
-          <div className="text-center py-8 text-muted-foreground">
-            <Shield className="w-10 h-10 mx-auto mb-3 opacity-40" />
-            <p className="text-sm font-medium mb-1">No clusters have OPA Gatekeeper installed</p>
-            <p className="text-xs mb-4">Install Gatekeeper on a cluster first, then create policies.</p>
-            <button
-              onClick={() => {
-                onClose()
-                const firstCluster = Object.keys(statuses)[0]
-                if (firstCluster) {
-                  startMission({
-                    title: `Install OPA Gatekeeper`,
-                    description: 'Set up OPA Gatekeeper for policy enforcement',
-                    type: 'deploy',
-                    cluster: firstCluster,
-                    initialPrompt: `I want to install OPA Gatekeeper on my cluster. Please help me install it using the official Helm chart and verify the installation is working.`,
-                    context: {},
-                  })
-                }
-              }}
-              disabled={Object.keys(statuses).length === 0}
-              className="px-4 py-2 bg-purple-500/20 text-purple-400 rounded-lg hover:bg-purple-500/30 transition-colors text-sm disabled:opacity-50"
-            >
-              Install Gatekeeper with AI
-            </button>
-          </div>
-        ) : (
-          <div className="space-y-4">
-            {/* Target Cluster Selector */}
-            <div>
-              <label className="block text-xs text-muted-foreground mb-1.5">Target Cluster</label>
-              <select
-                value={selectedCluster}
-                onChange={(e) => setSelectedCluster(e.target.value)}
-                disabled={isAnalyzing}
-                className="w-full px-3 py-2 bg-secondary/50 border border-border rounded-lg text-sm text-foreground focus:outline-none focus:ring-1 focus:ring-purple-500/50 disabled:opacity-50"
-              >
-                {installedClusters.map(name => (
-                  <option key={name} value={name}>{name}</option>
-                ))}
-              </select>
-            </div>
-
-            {/* Flow: Choose creation method */}
-            {flow === 'choose' && (
-              <div className="space-y-2">
-                <p className="text-xs text-muted-foreground font-medium border-b border-border/50 pb-1">
-                  How would you like to create a policy?
-                </p>
-
-                {/* Analyze & Suggest */}
-                <button
-                  onClick={handleAnalyzeAndSuggest}
-                  disabled={!selectedCluster || isAnalyzing}
-                  className="w-full p-3 rounded-lg bg-secondary/30 hover:bg-purple-500/10 border border-transparent hover:border-purple-500/30 transition-all text-left group disabled:opacity-50"
-                >
-                  <div className="flex items-start gap-3">
-                    <div className="mt-0.5">
-                      {isAnalyzing
-                        ? <Loader2 className="w-5 h-5 text-purple-400 animate-spin" />
-                        : <ScanSearch className="w-5 h-5 text-purple-400" />
-                      }
-                    </div>
-                    <div>
-                      <p className="text-sm font-medium text-foreground group-hover:text-purple-400 transition-colors">
-                        {isAnalyzing ? 'Analyzing cluster...' : 'Analyze Cluster & Suggest Policies'}
-                      </p>
-                      <p className="text-xs text-muted-foreground mt-0.5">
-                        AI scans security issues, missing limits, privileged containers and suggests OPA policies to prevent them
-                      </p>
-                    </div>
-                  </div>
-                </button>
-
-                {/* Describe What You Need */}
-                <button
-                  onClick={() => setFlow('describe')}
-                  disabled={!selectedCluster}
-                  className="w-full p-3 rounded-lg bg-secondary/30 hover:bg-purple-500/10 border border-transparent hover:border-purple-500/30 transition-all text-left group disabled:opacity-50"
-                >
-                  <div className="flex items-start gap-3">
-                    <MessageSquareText className="w-5 h-5 text-blue-400 mt-0.5" />
-                    <div>
-                      <p className="text-sm font-medium text-foreground group-hover:text-purple-400 transition-colors">
-                        Describe What You Need
-                      </p>
-                      <p className="text-xs text-muted-foreground mt-0.5">
-                        Tell AI what you want to enforce and it will generate the policy YAML
-                      </p>
-                    </div>
-                  </div>
-                </button>
-
-                {/* From Template */}
-                <button
-                  onClick={() => setFlow('template')}
-                  disabled={!selectedCluster}
-                  className="w-full p-3 rounded-lg bg-secondary/30 hover:bg-secondary/50 border border-transparent hover:border-border transition-all text-left group disabled:opacity-50"
-                >
-                  <div className="flex items-start gap-3">
-                    <LayoutTemplate className="w-5 h-5 text-green-400 mt-0.5" />
-                    <div>
-                      <p className="text-sm font-medium text-foreground group-hover:text-purple-400 transition-colors">
-                        From Template
-                      </p>
-                      <p className="text-xs text-muted-foreground mt-0.5">
-                        Start from a pre-built policy template
-                      </p>
-                    </div>
-                  </div>
-                </button>
-
-                {/* Custom YAML */}
-                <button
-                  onClick={() => setFlow('yaml')}
-                  disabled={!selectedCluster}
-                  className="w-full p-3 rounded-lg bg-secondary/30 hover:bg-secondary/50 border border-transparent hover:border-border transition-all text-left group disabled:opacity-50"
-                >
-                  <div className="flex items-start gap-3">
-                    <FileCode className="w-5 h-5 text-amber-400 mt-0.5" />
-                    <div>
-                      <p className="text-sm font-medium text-foreground group-hover:text-purple-400 transition-colors">
-                        Custom YAML
-                      </p>
-                      <p className="text-xs text-muted-foreground mt-0.5">
-                        Write or paste ConstraintTemplate + Constraint YAML directly
-                      </p>
-                    </div>
-                  </div>
-                </button>
-              </div>
-            )}
-
-            {/* Flow: Describe what you need */}
-            {flow === 'describe' && (
-              <div className="space-y-3">
-                <p className="text-xs text-muted-foreground">
-                  Describe the policy you want in plain English. AI will generate the ConstraintTemplate and Constraint YAML.
-                </p>
-                <textarea
-                  value={userDescription}
-                  onChange={(e) => setUserDescription(e.target.value)}
-                  className="w-full h-32 p-3 bg-secondary/50 border border-border rounded-lg text-sm text-foreground resize-none focus:outline-none focus:ring-1 focus:ring-purple-500/50"
-                  placeholder="e.g., Block all pods that don't have a 'team' label, require all containers to have memory limits, prevent images from untrusted registries..."
-                  autoFocus
-                />
-                <div className="flex justify-end">
-                  <button
-                    onClick={handleDescribeMission}
-                    disabled={!userDescription.trim()}
-                    className="flex items-center gap-2 px-4 py-2 bg-purple-500 text-white rounded-lg hover:bg-purple-600 disabled:opacity-50 disabled:cursor-not-allowed transition-colors text-sm"
-                  >
-                    <Sparkles className="w-4 h-4" />
-                    Generate with AI
-                  </button>
-                </div>
-              </div>
-            )}
-
-            {/* Flow: From Template */}
-            {flow === 'template' && (
-              <div className="space-y-2">
-                <p className="text-xs text-muted-foreground mb-2">
-                  Choose a template. You can edit the YAML before applying.
-                </p>
-                {POLICY_TEMPLATES.map(template => (
-                  <button
-                    key={template.name}
-                    onClick={() => handleUseTemplate(template)}
-                    className="w-full p-3 rounded-lg bg-secondary/30 hover:bg-secondary/50 transition-colors text-left"
-                  >
-                    <div className="flex items-center justify-between mb-1">
-                      <span className="text-sm font-medium text-foreground">{template.name}</span>
-                      <span className="text-xs text-muted-foreground">{template.kind}</span>
-                    </div>
-                    <p className="text-xs text-muted-foreground">{template.description}</p>
-                  </button>
-                ))}
-              </div>
-            )}
-
-            {/* Flow: Custom YAML / Template editor */}
-            {flow === 'yaml' && (
-              <div className="space-y-3">
-                <div className="flex items-center justify-between text-xs">
-                  <span className="text-muted-foreground">
-                    YAML will be applied to: <span className="text-foreground">{selectedCluster}</span>
-                  </span>
-                  <button
-                    onClick={() => {
-                      navigator.clipboard.writeText(yamlContent)
-                      showToast('Copied to clipboard', 'success')
-                    }}
-                    className="flex items-center gap-1 text-muted-foreground hover:text-foreground transition-colors"
-                  >
-                    <Copy className="w-3 h-3" />
-                    Copy
-                  </button>
-                </div>
-                <textarea
-                  value={yamlContent}
-                  onChange={(e) => setYamlContent(e.target.value)}
-                  className="w-full h-[40vh] p-3 bg-secondary/50 border border-border rounded-lg font-mono text-sm text-foreground resize-none focus:outline-none focus:ring-1 focus:ring-purple-500/50"
-                  placeholder="# Paste or write your ConstraintTemplate and Constraint YAML here..."
-                  spellCheck={false}
-                  autoFocus
-                />
-                <div className="flex justify-end">
-                  <button
-                    onClick={handleApplyCustomYaml}
-                    disabled={!yamlContent.trim()}
-                    className="px-4 py-2 bg-purple-500 text-white rounded-lg hover:bg-purple-600 disabled:opacity-50 disabled:cursor-not-allowed transition-colors text-sm"
-                  >
-                    Apply with AI
-                  </button>
-                </div>
-              </div>
-            )}
-          </div>
-        )}
-      </BaseModal.Content>
-
-      <BaseModal.Footer>
-        <button
-          onClick={onClose}
-          className="px-4 py-2 text-muted-foreground hover:text-foreground transition-colors text-sm"
-        >
-          Cancel
-        </button>
-        <div className="flex-1" />
-      </BaseModal.Footer>
-    </BaseModal>
-  )
 }
 
 // Sort comparators that use statuses lookup via closure
